Small cleanup of IRingRange.
