+++ test/TesterInternal/General/RingTests_Standalone.cs
@@ -216,7 +216,7 @@ namespace UnitTests.General
         public RangeBreakable()
         {
             ranges = new List<SingleRange>(1);
-            ranges.Add(new SingleRange(0, 0));
+            ranges.Add(RangeFactory.CreateFullRange() as SingleRange);
         }
 
         public bool Remove(IRingRange range)
@@ -233,7 +233,7 @@ namespace UnitTests.General
                         ranges.Remove(m);
                         if (s.Begin != s.End) // if s is full range as well, then end of story ... whole range is covered
                         {
-                            ranges.Add(new SingleRange(s.End, s.Begin));
+                            ranges.Add(RangeFactory.CreateRange(s.End, s.Begin) as SingleRange);
                         }
                         break;
                     }
@@ -244,11 +244,11 @@ namespace UnitTests.General
                         ranges.Remove(m);
                         if (s.Begin != m.Begin)
                         {
-                            ranges.Add(new SingleRange(m.Begin, s.Begin));
+                            ranges.Add(RangeFactory.CreateRange(m.Begin, s.Begin) as SingleRange);
                         }
                         if (s.End != m.End)
                         {
-                            ranges.Add(new SingleRange(s.End, m.End));
+                            ranges.Add(RangeFactory.CreateRange(s.End, m.End) as SingleRange);
                         }
                         break;
                     }
